Solving a bug
When a user adds a comment,the offline comment displays with it's uuid and POST request is made to the server. What if a user wants to delete his comment in the database with the id it has in the db but not uuid? There comes the bug. MYSQL throws an error saying: Cannot delete comment with id=uuid or 3453&&&6665544 whilst the id may be 32 in the db. So we then execute the fetch comments function as user adds comment